[quantization] Unroll all heads (#490)
This commit unrolls all heads in llama wrapper.

TICO-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>

+++ tico/quantization/wrapq/examples/llama/quantize_llama_decoder_layer.py
@@ -44,7 +44,7 @@ from tico.utils.utils import SuppressWarning
 MODEL_NAME = "Maykeye/TinyLLama-v0"
 MAX_S = 256
 
-model = AutoModelForCausalLM.from_pretrained(MODEL_NAME)
+model = AutoModelForCausalLM.from_pretrained(MODEL_NAME, dtype="float32")
 tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME)
 
 # Make sure pad token exists (Llama often uses eos as pad)
